do some fancy magic stuff
