--- tests/setup-env-node.js
@@ -1,3 +1,0 @@
-// Make sure that our tests do the use fakeTimer mock:
-// https://github.com/testing-library/dom-testing-library/issues/300
-global.useFakeTimers = true
